+++ src/LawnDart.Messaging/TypeForwards.cs
@@ -1,0 +1,4 @@
+using System.Runtime.CompilerServices;
+using LawnDart.Messaging;
+
+[assembly: TypeForwardedTo(typeof(ICommandDispatcher))]